Revert "Added UserDetails, Patient and Receptionist entities with relationships"
This reverts commit 3c911a11157b480fcf072bcb6ebad3415ba34eac.

+++ src/main/java/com/example/HospitalManagementSystem/repository/UserDetailsRepository.java
@@ -1,7 +1,9 @@
 package com.example.HospitalManagementSystem.repository;
 
-import com.example.HospitalManagementSystem.model.UserDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserDetailsRepository extends JpaRepository<UserDetails, Long> {
+import com.example.HospitalManagementSystem.model.UserDetails;
+
+public interface UserDetailsRepository extends JpaRepository<UserDetails, Integer> {
+
 }
